Work around NIPM/libsolv issue
libsolv (which NIPM uses) has an issue where recommended packages are not listed as upgradable when the version of the recommendation changes: https://github.com/openSUSE/libsolv/issues/387

This is preventing the newer versions of custom devices from showing up as updates in NIPM; instead, only the top level item shows up as an update. Following a discussion with the NIPM team, we believe recommending the previous version will still result in NIPM prompting for an update (since it always recommends the latest), while working around this issue.

We may need to do some work in the future to explicitly not include these old versions in the feed, since the internal feedbuild tooling by default matches the version of the recommendation.