Virtual package should require VeriStand
The virtual package should require at least _some_ version of VeriStand be installed.

Need to use substitution variables for 2017 and 2018, since these are actually upgrade codes and not package names. Since 2019, a VeriStand package (technically e.g. `ni-veristand-2019-core-x86`, not `ni-veristand-2019`) provides `ni-veristand` with the corresponding version, so we can depend on this generically.